% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grmCAT.R
\name{grmCAT}
\alias{grmCAT}
\title{MTB Graded Response Model Engine}
\usage{
grmCAT(
  ipar,
  resp,
  minTheta = -6,
  maxTheta = 6,
  nQpts = 121,
  prevTheta = NULL,
  prevSD = NULL,
  muEAP = 0,
  sdEAP = 3,
  scoreCumulativeDiffuse = TRUE,
  start_item = 0,
  minNI = 4,
  maxSD = 0.3,
  maxNI = 8,
  deltaSD = 0.01
)
}
\arguments{
\item{ipar}{An item parameter matrix in slope/threshold format. Each item is on a separate line.
The column names should be "a" and "CB1" through "CB#" for
the slope and relevant thresholds.}

\item{resp}{Response matrix (number of rows = number of participants and number of columns = number of items)}

\item{minTheta}{Minimum theta value (PROMIS default is -4)}

\item{maxTheta}{Maximum theta value (PROMIS default is 4)}

\item{nQpts}{number of quadrature points, with min minTheta and max maxTheta}

\item{prevTheta}{estimated theta from prior CAT administration}

\item{prevSD}{estimated SD from prior CAT administration}

\item{muEAP}{Estimated A Posterior mean. Defaults to NULL.}

\item{sdEAP}{Estimated A Posterior SD. Defaults to NULL.}

\item{scoreCumulativeDiffuse}{Logical to indicate whether the theta grid should be modified prior to administering the first item based
on the final score estimate from the previous administration. Defaults to TRUE.}

\item{start_item}{First item to start the CAT engine.}

\item{minNI}{Minimum number of items to administer. Defaults to 4.}

\item{maxSD}{Maximum SD. Defaults to 0.3.}

\item{maxNI}{Maximum number of items to administer. Defaults to 8.}

\item{deltaSD}{Mininum change for SD.}
}
\value{
This function returns a list with five elements. The first two elements are scalar elements with
the final theta estimate and the standard deviation of the score elements.The third element is a dataframe
with the item administration history. The first row is the prior. The second row through the final are
the items administered, the chosen response category score, and the theta/sd history. The fourth and fifth
elements are the probability differences array and the item information matrix,
both of which are calculated by the the matrixInfoGRM function.
}
\description{
This function is a wrapper to simulate the graded response model engine programmed within MobileToolbox.
It utilizes matrixInfoGRM, selectMPWI, and calcEAP funtions for item selection and scoring.
}
\examples{
# define item parameter bank; number of rows is the number of columns
# a = item slope; CB1 to CB# = category bins/thresholds for response; NCAT = number of categories for item

ipar <- data.frame(a = c(2.99, 3.53,  3.10),
                  CB1 = c(-1.18, -2.79,  -2.12),
                  CB2 = c(-0.499, -1.85,  -1.27),
                  CB3 = c(0.17, -1.19,  -0.52),
                  CB4 = c(0.653, -0.551,  0.036),
                  NCAT = 5)
#simulate responses
set.seed(22416)
# 20 participants and their responses from the three items
resp <- round(matrix(stats::runif(20*3, 1, 5), 20, 3))

# Currently grmCAT is NOT vectorized and only 1 participant can be run at a time (unless using apply)

# run
grmCAT_output <- grmCAT(ipar, resp[,1],minTheta=-6,maxTheta=6,nQpts=121,prevTheta=NULL,
prevSD=NULL,muEAP=0,sdEAP=3,scoreCumulativeDiffuse=TRUE,
start_item=0,minNI=4,maxSD=0.3,maxNI=8,deltaSD=0.01)

# set scoreCumulativeDiffuse = T and prevTheta = 1
grmCAT_sCD_true <- grmCAT(ipar, resp[,1],minTheta=-6,maxTheta=6,nQpts=121,prevTheta=1,
prevSD=NULL,muEAP=0,sdEAP=3,scoreCumulativeDiffuse=TRUE,
start_item=0,minNI=4,maxSD=0.3,maxNI=8,deltaSD=0.01)

grmCAT_sCD_false <- grmCAT(ipar, resp[,1],minTheta=-6,maxTheta=6,nQpts=121,prevTheta=1,
prevSD=NULL,muEAP=0,sdEAP=3,scoreCumulativeDiffuse=FALSE,
start_item=0,minNI=4,maxSD=0.3,maxNI=8,deltaSD=0.01)

# changing the scoreCumulativeDiffuse option does not change item selection with only three items
# theta estimates should generally be very close, if not identical
 grmCAT_sCD_true$finalTheta; grmCAT_sCD_false$finalTheta

# or with apply
#' allResp <- apply(resp,1, function(y) grmCAT(ipar, resp=y))


 #### Using the PROMIS Pediatric Strength Impact item bank
 set.seed(542)
# 5 participants with random responses to all 12 items
data(iparPedImp)
respSI <- round(matrix(stats::runif(12*5, 1, 5), 5, 12))
grm_out1 <- apply(respSI,1,function(y) grmCAT(ipar=iparPedImp[,2:7], resp=y,
minTheta=-4, maxTheta=4, nQpts=81,minNI=5,maxNI=12)) # PROMIS Peds defaults

table(sapply(grm_out1, function(y) nrow(y$scoreHistory)-1)) # all participants got 5 items
table(sapply(grm_out1, function(y) y$scoreHistory$itemAdminHistory))

}
\seealso{
Other GRM_CAT functions: 
\code{\link{calcEAP}()},
\code{\link{matrixInfoGRM}()},
\code{\link{selectMPWI}()}
}
\concept{GRM_CAT functions}
