% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iparDep}
\alias{iparDep}
\title{Item Parameters for the PROMIS Adult Depression Item Bank}
\format{
\describe{
 \item{itemID}{Item Identifier}
 \item{a}{slope}
 \item{CB1}{Threshold (or category boundary) 1}
 \item{CB2}{Threshold (or category boundary) 2}
 \item{CB3}{Threshold (or category boundary) 3}
 \item{CB4}{Threshold (or category boundary) 4}
 \item{NCAT}{Number of item categories used by this item}
}
}
\usage{
iparDep
}
\description{
A dataset of IRT calibrations using slope-threshold parameterization for the graded response model. There are 15 items with 5 response categories each.
}
\keyword{datasets}
