% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_xPL.R
\name{maxInfo_xPL}
\alias{maxInfo_xPL}
\title{Dichotomous Max Info}
\usage{
maxInfo_xPL(iparFull, targTheta)
}
\arguments{
\item{iparFull}{A matrix with rows for items and three columns. The first column is the slope,
the second is the intercept (not the threshold),
and the third is the guessing parameter (on the original scale, not on the logit scale).}

\item{targTheta}{The target theta at which the item information is to be calculated and the next item chosen.}
}
\value{
This function returns a vector of integers indicating row numbers from iparFull with the maximum information provided at the target theta value.
}
\description{
This function takes parameters from a dichotomous item bank and a target theta estimate.
It calculates the item information at the target theta values (using the xPL_TIF function),
and organizes the items from maximum to minimum information provided.
}
\seealso{
Other dichotomous functions: 
\code{\link{dichotomousEngine}()},
\code{\link{xPL_MLE}()},
\code{\link{xPL_TIF}()}
}
\concept{dichotomous functions}
