% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectMPWI.R
\name{selectMPWI}
\alias{selectMPWI}
\title{Maximum Posterior Weighted Information (MPWI) Item Selection}
\usage{
selectMPWI(matrixInfo, currentProbGrid, nItems)
}
\arguments{
\item{matrixInfo}{This is the information matrix for the item bank calulated using the matrixInfoGRM() function.
The information matrix should have been called by the GRM engine prior to item selection. But if used alone,
this will need to be applied.}

\item{currentProbGrid}{The probability grid must be supplied to weight the item information matrix.
It may be the normal density (e.g., if no items have been calculated), or  it may be the density of the current EAP score estimate.}

\item{nItems}{The number of items in the item bank.}
}
\value{
After running MPWI item selection, a vector is returned equal in length to the number of items in the item bank.
The values are item information (across all categories) weighted by the current probability grid.
}
\description{
This function is utilized by the GRM engine to select the next item using the MPWI
item selection method. It is called within the engine, but can be used independently, if so desired.
}
\seealso{
matrixInfoGRM
}
