% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spellPars}
\alias{spellPars}
\title{Item Parameters for the Dichotomous Mobile Toolbox Spelling Item Bank}
\format{
\describe{
  \item{stepID}{Step Identifier}
  \item{a}{Slope}
  \item{d}{Intercept}
  \item{g}{Pseudo-guessing}
  \item{Experimental}{Indicator for Experimental Status}
  \item{calib}{An alternative method to represent experimental status where 0=experimental and 1=calibrated}
}
}
\usage{
spellPars
}
\description{
A dataset of IRT calibrations using slope-intercept parameterization for the 2-parameter logistic model. There are 498 items scored correct/incorrect, several of which are experimental items.
}
\keyword{datasets}
