% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE_xPL.R
\name{xPL_MLE}
\alias{xPL_MLE}
\title{Dichotmous MLE}
\usage{
xPL_MLE(
  ipar,
  u,
  st_th = 0,
  crit = 0.001,
  maxIter = 100,
  minTheta = -10,
  maxTheta = 10
)
}
\arguments{
\item{ipar}{A matrix with rows for items and three columns. The first column is the slope,
the second is the intercept (not the threshold),
and the third is the guessing parameter (on the original scale, not on the logit scale).}

\item{u}{A vector of response strings, coded 0 for incorrect and 1 for correct.
The length of this vector must be equal to the number of rows (i.e., the number of items) in ipar.}

\item{st_th}{Starting theta to initialize the Newton Raphson method. Defaults to 0.}

\item{crit}{Convergence criteria for Newton Raphson method. Defaults to 0.001.}

\item{maxIter}{Maximum number of iterations for the Newton Raphson method. Defaults to 100.}

\item{minTheta}{Minimum permissible theta value. Defaults to -10.}

\item{maxTheta}{Maximum permissible theta value. Defaults to +10.}
}
\value{
Returns a data.frame with the theta estimate at convergence (or minTheta or maxTheta if convergence was not met),
and the standard error associated with that estimate (which is Inf if not converged),
and the number of iterations taken by the Newton Raphson algorithm to reach convergence (or maxIter if not converged).
}
\description{
This function calculates the maximum likelihood estimate for dichotomous items. It is based off of the code provided in Baker and Kim 2017.
However, it includes modifications for slope/intercept (as opposed to slope/threshold) parameterization,
though it maintains original-scale guessing parameterization. It supports the Rasch/1PL, 2PL, and 3PL models provided that
they are appropriately parameterized for this function.
}
