% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIF_xPL.R
\name{xPL_TIF}
\alias{xPL_TIF}
\title{Dichotomous TIF}
\usage{
xPL_TIF(ipar, thetaGrid = seq(-4.5, 4.5, 0.1))
}
\arguments{
\item{ipar}{A matrix with rows for items and three columns. The first column is the slope,
the second is the intercept (not the threshold),
and the third is the guessing parameter (on the original scale, not on the logit scale).}

\item{thetaGrid}{A vector of theta estimates at which the test information will be calculated.
This defaults to a vector from -4.5 to 4.5, incremented by 0.1.}
}
\value{
This function returns a list. The first element is a matrix with the item information.
The thetaGrid forms the rows, and the number of items forms the columns.
The second element of the list is a vector the same length as thetaGrid. It is the test information function estimated at each point along the thetaGrid.
}
\description{
This function calculates the test information function for an item bank.
It supports the Rasch/1PL, 2PL, and 3PL models provided that they are appropriately parameterized for this function.
}
