% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIF_xPL.R
\name{xPL_TIF}
\alias{xPL_TIF}
\title{Dichotomous TIF}
\usage{
xPL_TIF(ipar, thetaGrid = seq(-4.5, 4.5, 0.1))
}
\arguments{
\item{ipar}{A matrix with rows for items and three columns. The first column is the slope,
the second is the intercept (not the threshold),
and the third is the guessing parameter (on the original scale, not on the logit scale).}

\item{thetaGrid}{A vector of theta estimates at which the test information will be calculated.
This defaults to a vector from -4.5 to 4.5, incremented by 0.1.}
}
\description{
This function calculates the test information function for an item bank
}
